backend/requirements.txt
fastapi
uvicorn[standard]
osmnx
networkx
requests
python-dotenv
shapely
scikit-learn
